Remove nnx e2e smoke test; inline its tiny-model builder
test_system_e2e.py was a random-weight construction/streaming smoke (no
jax<->nnx parity). Its sole surviving value was the _build_tiny_system
helper, which tests/nnx/test_system.py already depended on; move that
builder into test_system.py and drop the redundant smoke tests.
